pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "helixar-sentinel"
version = "0.1.0"
description = "MCP security scanner by Helixar"
readme = "README.md"
license = { file = "LICENSE" }
authors = [{ name = "Helixar Security Research", email = "security@helixar.ai" }]
keywords = ["mcp", "security", "scanner", "llm", "ai-security"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: Software Development :: Quality Assurance",
]
requires-python = ">=3.9"
dependencies = [
    "click>=8.1",
    "rich>=13.0",
    "requests>=2.31",
    "pyyaml>=6.0",
    "docker>=7.0",
    "cryptography>=41.0",
    "jinja2>=3.1",
    "jsonschema>=4.20",
    "urllib3>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "pytest-mock>=3.12",
    "responses>=0.24",
    "ruff>=0.1",
    "mypy>=1.7",
]

[project.scripts]
sentinel = "sentinel.cli:main"

[project.urls]
Homepage = "https://github.com/Helixar-AI/sentinel"
Repository = "https://github.com/Helixar-AI/sentinel"
"Bug Tracker" = "https://github.com/Helixar-AI/sentinel/issues"
Documentation = "https://github.com/Helixar-AI/sentinel#readme"

[tool.setuptools.packages.find]
where = ["."]
include = ["sentinel*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
